chore: rename and reordering of the folders
