+++ src/components/core/EaNavButton.tsx
