--- src/global-components/file.txt
@@ -1,1 +1,0 @@
-qui vanno le component globali
