+++ src/hooks/core/useNavBarEAHook.tsx
