--- src/pages/file.txt
@@ -1,1 +1,0 @@
-qui vanno le pagine
